Update Marcas y Productos CRUD, mejorando aspectos visuales con Bootstrap

+++ eCommerceMVC/Controllers/MarcasController.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using eCommerceMVC.Models;
 
@@ -27,108 +25,72 @@ namespace eCommerceMVC.Controllers
         // GET: Marcas/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var marca = await _context.Marcas
                 .FirstOrDefaultAsync(m => m.IdMarca == id);
-            if (marca == null)
-            {
-                return NotFound();
-            }
+            if (marca == null) return NotFound();
 
             return View(marca);
         }
 
         // GET: Marcas/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        public IActionResult Create() => View();
 
         // POST: Marcas/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("IdMarca,Descripcion,Activo,FechaRegistro")] Marca marca)
+        public async Task<IActionResult> Create(Marca marca)
         {
-            if (ModelState.IsValid)
-            {
-                _context.Add(marca);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(marca);
+            if (!ModelState.IsValid) return View(marca);
+
+            marca.FechaRegistro = DateTime.Now; // Fecha automática
+
+            _context.Add(marca);
+            await _context.SaveChangesAsync();
+
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Marcas/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var marca = await _context.Marcas.FindAsync(id);
-            if (marca == null)
-            {
-                return NotFound();
-            }
+            if (marca == null) return NotFound();
+
             return View(marca);
         }
 
         // POST: Marcas/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("IdMarca,Descripcion,Activo,FechaRegistro")] Marca marca)
+        public async Task<IActionResult> Edit(int id, Marca marca)
         {
-            if (id != marca.IdMarca)
-            {
-                return NotFound();
-            }
+            if (id != marca.IdMarca) return NotFound();
+            if (!ModelState.IsValid) return View(marca);
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(marca);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!MarcaExists(marca.IdMarca))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(marca);
+            var marcaExistente = await _context.Marcas.FindAsync(id);
+            if (marcaExistente == null) return NotFound();
+
+            // Actualizamos solo campos editables
+            marcaExistente.Descripcion = marca.Descripcion;
+            marcaExistente.Activo = marca.Activo;
+
+            await _context.SaveChangesAsync();
+
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Marcas/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var marca = await _context.Marcas
                 .FirstOrDefaultAsync(m => m.IdMarca == id);
-            if (marca == null)
-            {
-                return NotFound();
-            }
+            if (marca == null) return NotFound();
 
             return View(marca);
         }
@@ -139,18 +101,10 @@ namespace eCommerceMVC.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var marca = await _context.Marcas.FindAsync(id);
-            if (marca != null)
-            {
-                _context.Marcas.Remove(marca);
-            }
+            if (marca != null) _context.Marcas.Remove(marca);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool MarcaExists(int id)
-        {
-            return _context.Marcas.Any(e => e.IdMarca == id);
         }
     }
 }
